Rename account identity and classify permission failures
