+++ CLAUDE.md
@@ -81,7 +81,7 @@ Three things here are load-bearing:
 Three front-ends fill one file, `~/.delta-exchange-mcp/config.env`:
 
 - `store.py` owns the file — `path/read/ensure/write/insecure_permissions`. `ensure` creates it `0600` from a commented `TEMPLATE` on first run; `write` goes through dotenv's `set_key` so comments and unrelated settings survive. `config.setting(name)` resolves the process environment first and this file second, with empty meaning unanswered (a bundle substitutes every declared variable whether or not the field was filled).
-- `identity.py` owns the API-key account identity: `fetch_account_identity(client)` validates `result.user_id` from `/users/trading_preferences`. Credential checks, the account tool, and close-all resolution share it so the endpoint and response contract cannot drift.
+- `account_identity.py` owns the API-key account identity: `fetch_account_identity(client)` validates `result.user_id` from `/users/trading_preferences`. Credential checks, the account tool, and close-all resolution share it so the endpoint and response contract cannot drift.
 - `credentials.py` is the shared domain: `check(env, key, secret)` verifies the key through the identity helper, and `save(...)` writes the key, secret and environment together. Neither front-end owns these. `Check.code` carries Delta's own error code beside the rendered message so a caller can branch on which failure it was without matching on that message's text.
 - `store.write()` holds an OS-backed advisory lock around its complete copy-modify-replace transaction. This matters now that different clients can save disjoint scoped mode keys: without serialization, a stale staging copy can undo another client's successful trade-to-read de-escalation. The hidden lock file is persistent by design and contains no settings; the kernel releases its lock after a crash.
 - `credentials.overridden_by_client()` names the settings in the shared file that the process environment is beating, and both front-ends report it — `login` as a note on stderr, the form as an `overridden` status. Without it a save is silently useless: `config` resolves the environment first, so a client passing its own key (the bundle's `user_config`, VS Code's `inputs`, an edited Cursor entry) wins on every launch, and the form would verify one account, name it, and leave the server signing with another. **It compares what `config` resolves against what the file holds, never mere presence** — the Cursor install link sets `DELTA_MCP_ENV` for everyone, so a presence test would tell every Cursor user their working key was ignored. Regression test: `test_a_client_pinning_the_same_environment_is_not_reported`.
